UI fixes

+++ FrontEnd/frmCustomer.cs
@@ -11,6 +11,9 @@ namespace SecretCellar
     public partial class frmCustomer : Form
     {
         private Transaction _importedTransaction = null;
+        private string _lastSortedColumnName = "";
+
+
 
         public frmCustomer(Transaction transaction)
         {
@@ -149,47 +152,47 @@ namespace SecretCellar
         {
             if (customer_data_grid.SelectedRows.Count > 0)
             {
-                Customer i = DataAccess.instance.GetCustomer().FirstOrDefault(x => x.LastName == txt_lname.Text.Trim() && x.FirstName == txt_fname.Text.Trim());
+                if (frmManagerOverride.DidOverride("Create New Customer")) {
+                    Customer i = DataAccess.instance.GetCustomer().FirstOrDefault(x => x.LastName == txt_lname.Text.Trim() && x.FirstName == txt_fname.Text.Trim());
 
-                if (i != null)
-                {
-                    if(MessageBox.Show($@"Customer already exists.  Would you like to edit this customer?
+                    if (i != null) {
+                        if (MessageBox.Show($@"Customer already exists.  Would you like to edit this customer?
 Last Name: {i.LastName}
 First Name: {i.FirstName}
 E-mail: {i.Email}
 BusinessName: {i.BusinessName}
 Address: 
-                    {i.Address1}{(string.IsNullOrWhiteSpace ( i.Address2) ? "": $",{i.Address2}")} 
-  {i.City}, {i.State} {i.ZipCode}","Customer Exists", MessageBoxButtons.YesNo) == DialogResult.Yes)
-                    {
-                        SelectCustomer(i);
-                        return;
-                    }
-                }
-                else {
-                    if (txt_lname.Text.Replace(" ", "") != "" && txt_fname.Text.Replace(" ", "") != "") {
-                        i = new Customer {
-                            LastName = txt_lname.Text.Trim(),
-                            FirstName = txt_fname.Text.Trim(),
-                            Email = txt_email.Text.Trim(),
-                            BusinessName = txt_company.Text.Trim(),
-                            IsWholesale = cbo_wholesale.SelectedIndex != 0,
-                            CustomerDiscount = double.TryParse(txt_custDisc.Text, out double disc) ? disc : 0,
-                            PhoneNumber = txt_phone.Text,
-                            Address1 = txt_addr1.Text.Trim(),
-                            Address2 = txt_addr2.Text.Trim(),
-                            City = txt_city.Text.Trim(),
-                            State = (string)cmb_State.SelectedValue,
-                            ZipCode = txt_zip.Text.Trim()
-                        };
-
-                        i.CustomerID = DataAccess.instance.NewCustomer(i);
-                        refresh();
+                    {i.Address1}{(string.IsNullOrWhiteSpace(i.Address2) ? "" : $",{i.Address2}")} 
+  {i.City}, {i.State} {i.ZipCode}", "Customer Exists", MessageBoxButtons.YesNo) == DialogResult.Yes) {
+                            SelectCustomer(i);
+                            return;
+                        }
                     }
                     else {
-                        MessageBox.Show("First and Last name cannot be empty.", "Error");
-					}
-				}
+                        if (txt_lname.Text.Replace(" ", "") != "" && txt_fname.Text.Replace(" ", "") != "") {
+                            i = new Customer {
+                                LastName = txt_lname.Text.Trim(),
+                                FirstName = txt_fname.Text.Trim(),
+                                Email = txt_email.Text.Trim(),
+                                BusinessName = txt_company.Text.Trim(),
+                                IsWholesale = cbo_wholesale.SelectedIndex != 0,
+                                CustomerDiscount = double.TryParse(txt_custDisc.Text, out double disc) ? disc : 0,
+                                PhoneNumber = txt_phone.Text,
+                                Address1 = txt_addr1.Text.Trim(),
+                                Address2 = txt_addr2.Text.Trim(),
+                                City = txt_city.Text.Trim(),
+                                State = (string)cmb_State.SelectedValue,
+                                ZipCode = txt_zip.Text.Trim()
+                            };
+
+                            i.CustomerID = DataAccess.instance.NewCustomer(i);
+                            refresh();
+                        }
+                        else {
+                            MessageBox.Show("First and Last name cannot be empty.", "Error");
+                        }
+                    }
+                }
             }
         }
 
@@ -198,39 +201,41 @@ Address:
             if (customer_data_grid.SelectedRows.Count > 0)
             {
                 if (txt_lname.Text.Replace(" ", "") != "" && txt_fname.Text.Replace(" ", "") != "") {
-                    Customer i = DataAccess.instance.GetCustomer(uint.Parse(customer_data_grid.SelectedRows[0].Cells["customerID"].Value.ToString()));
+                    if (frmManagerOverride.DidOverride("Update Customer")) {
+                        Customer i = DataAccess.instance.GetCustomer(uint.Parse(customer_data_grid.SelectedRows[0].Cells["customerID"].Value.ToString()));
 
-                    i.LastName = txt_lname.Text;
-                    i.FirstName = txt_fname.Text;
-                    i.Email = txt_email.Text;
-                    i.BusinessName = txt_company.Text;
-                    // i.IsWholesale = (bool)cbo_wholesale.SelectedValue; following if statement does work
-                    if (cbo_wholesale.SelectedIndex == 1) {
-                        i.IsWholesale = true;
+                        i.LastName = txt_lname.Text;
+                        i.FirstName = txt_fname.Text;
+                        i.Email = txt_email.Text;
+                        i.BusinessName = txt_company.Text;
+                        // i.IsWholesale = (bool)cbo_wholesale.SelectedValue; following if statement does work
+                        if (cbo_wholesale.SelectedIndex == 1) {
+                            i.IsWholesale = true;
+                        }
+                        else {
+                            i.IsWholesale = false;
+                        }
+
+                        if (txt_custDisc.Text == null) {
+                            i.CustomerDiscount = 0;
+                        }
+                        else {
+                            i.CustomerDiscount = Convert.ToUInt32(txt_custDisc.Text);
+                        }
+
+                        i.PhoneNumber = txt_phone.Text;
+                        i.Address1 = txt_addr1.Text;
+                        i.Address2 = txt_addr2.Text;
+                        i.City = txt_city.Text;
+                        i.State = (string)cmb_State.SelectedValue;
+                        i.ZipCode = txt_zip.Text;
+
+
+                        i.CustomerID = DataAccess.instance.UpdateCustomer(i);
+                        //  customers.Add(i);
+                        //  dataAccess.UpdateCustomer(i);
+                        refresh();
                     }
-                    else {
-                        i.IsWholesale = false;
-                    }
-
-                    if (txt_custDisc.Text == null) {
-                        i.CustomerDiscount = 0;
-                    }
-                    else {
-                        i.CustomerDiscount = Convert.ToUInt32(txt_custDisc.Text);
-                    }
-
-                    i.PhoneNumber = txt_phone.Text;
-                    i.Address1 = txt_addr1.Text;
-                    i.Address2 = txt_addr2.Text;
-                    i.City = txt_city.Text;
-                    i.State = (string)cmb_State.SelectedValue;
-                    i.ZipCode = txt_zip.Text;
-
-
-                    i.CustomerID = DataAccess.instance.UpdateCustomer(i);
-                    //  customers.Add(i);
-                    //  dataAccess.UpdateCustomer(i);
-                    refresh();
                 }
                 else {
                     MessageBox.Show("First and Last name cannot be empty.", "Error");
@@ -280,23 +285,79 @@ Address:
 		private void button_UpdateCredit_Click(object sender, EventArgs e) {
 
             Customer customer = DataAccess.instance.GetCustomer(uint.Parse(customer_data_grid.SelectedRows[0].Cells["customerID"].Value.ToString()));
-
             frmCustomerCredit frmCustomerCredit = new frmCustomerCredit(customer);
-            DialogResult result = frmCustomerCredit.ShowDialog();
-
-            if (result.Equals(DialogResult.OK)) {
-                refresh();
-			}
+            frmCustomerCredit.ShowDialog();
+            refresh();
         }
 
 		private void button_delete_Click(object sender, EventArgs e) {
             if (customer_data_grid.SelectedRows.Count > 0) {
-                Customer customer = DataAccess.instance.GetCustomer(uint.Parse(customer_data_grid.SelectedRows[0].Cells["customerID"].Value.ToString()));
-
-                DataAccess.instance.DeleteCustomer(customer);
-
-                refresh();
+                if (frmManagerOverride.DidOverride("Delete Customer")) {
+                    Customer customer = DataAccess.instance.GetCustomer(uint.Parse(customer_data_grid.SelectedRows[0].Cells["customerID"].Value.ToString()));
+                    DataAccess.instance.DeleteCustomer(customer);
+                    refresh();
+                }
 			}
         }
-	}
+
+        private void customer_data_grid_ColumnHeaderMouseClick(object sender, DataGridViewCellMouseEventArgs e) {
+            string columnName = customer_data_grid.Columns[e.ColumnIndex].Name;
+            int currentColumn = customer_data_grid.FirstDisplayedScrollingColumnIndex;
+            List<Customer> customers = new List<Customer>();
+
+            foreach (DataGridViewRow row in customer_data_grid.Rows) {
+                if (uint.TryParse(row.Cells[0].Value.ToString(), out uint customerId)) {
+                    Customer customer = DataAccess.instance.GetCustomer(customerId);
+                    customers.Add(customer);
+                }
+            }
+
+            var customersFiltered = customers.Select(x => new {
+                customerID = x.CustomerID,
+                last_name = x.LastName,
+                first_name = x.FirstName,
+                phone = x.PhoneNumber,
+                email = x.Email,
+                business_name = x.BusinessName,
+                isWholesale = x.IsWholesale,
+                customerDiscount = x.CustomerDiscount,
+                addr1 = x.Address1,
+                addr2 = x.Address2,
+                city = x.City,
+                state = x.State,
+                zip = x.ZipCode,
+                Credit = x.Credit
+            });
+
+            if (columnName == _lastSortedColumnName) {
+                customersFiltered = customersFiltered.Reverse().ToList();
+            }
+            else {
+                customersFiltered = customersFiltered
+                .OrderBy(x => {
+                    switch (columnName) {
+                        case "last_name": { return x.last_name; }
+                        case "first_name": { return x.first_name; }
+                        case "phone": { return x.phone; }
+                        case "email": { return x.email; }
+                        case "Credit": { return x.Credit + ""; }
+                        case "business_name": { return x.business_name; }
+                        case "isWholesale": { return x.isWholesale + ""; }
+                        case "customer_discount": { return x.customerDiscount + ""; }
+                        case "addr1": { return x.addr1; }
+                        case "addr2": { return x.addr2; }
+                        case "city": { return x.city; }
+                        case "state": { return x.state; }
+                        case "zip": { return x.zip; }
+                        default: { return x.last_name; }
+                    }
+                })
+                .ToList();
+            }
+
+            customer_data_grid.DataSource = customersFiltered;
+            _lastSortedColumnName = columnName;
+            customer_data_grid.FirstDisplayedScrollingColumnIndex = currentColumn;
+        }
+    }
 }
